Changed the way plugins are loaded in windows – now when loading a plugin – windows will search for dependant dlls of the plugin also in the directory were the plugin is located. The helps for example in the case where you want to load Plugin_CgProgramManager from a directory that is not the application directory and not in the environment path – now you can place the cg.dll (a dependent dll) next to it.
The only case this change can create a problem – is if you have a plug-in in a different path then the application and an un-updated dependent dll next to it in that dir, this is an unrealistic case. I really don’t think this is a risky change.

I added this code also to 1.6